requirements.txt
# openai==0.27.0
# SQLAlchemy==2.0.4
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
# Editable install with no version control (gpt3-contextual==0.0.0)
-e c:\users\user\documents\個人開発\lazy-ai\gpt3-contextual-main
greenlet==3.0.3
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langsmith==0.1.59
line-bot-sdk==3.11.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
ngrok==1.3.0
numexpr==2.10.0
numpy==1.26.4
openai==0.27.0
openapi-schema-pydantic==1.2.4
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
protobuf==3.20.3
pyarrow==16.1.0
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.4
starlette==0.37.2
stellarstation==0.14.0
streamlit==1.34.0
streamlit-chat==0.1.1
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
watchdog==4.0.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4